upnpav: Move RendererDouble from Multimedia to UPnPAVTestHelper
